code improve: Extract the Table class
